Rename LegendreGaussLobatto.hpp/cpp to Legendre.hpp/cpp
