+++ src/NumericalAlgorithms/LinearOperators/DefiniteIntegral.cpp
@@ -6,7 +6,7 @@
 #include "DataStructures/DataVector.hpp"
 #include "DataStructures/Index.hpp"
 #include "ErrorHandling/Assert.hpp"
-#include "NumericalAlgorithms/Spectral/LegendreGaussLobatto.hpp"
+#include "NumericalAlgorithms/Spectral/Legendre.hpp"
 #include "Utilities/Blas.hpp"
 
 namespace {
